Buttons and Images
Arranged scripts, styles and resources to their respective folders.

Replaced the main font of the site to Poppins. Original font looked ugly.

Removed the non-functional MENU button from the navbar, replaced it with the contact button.

Reverted hover-action color of the DOWNLOAD CV button

Former image of the developer looked too irrelevant and messy on the page given its shiny out-of-the-palette colors, removed background from the original image and replaced with the new transparent one

Reduced social icons translation to 5 upon hover action